The Bathtub Model of Epidemiology
The Bathtub Model of Epidemiology: A Simple Analogy for a Complex Reality.
The Bathtub analogy is a widely used analogy that helps to visualize and understand the relationship between the incidence and prevalence of a disease within a population. It simplifies complex epidemiological concepts by comparing the burden of a disease to
the amount of water in a bathtub. This model is particularly useful for explaining how the number of existing cases is
influenced by the rate of new cases and the rate at which people either recover or die. By adding a simple modification, the model can also effectively illustrate the impact of disease recurrence.